+++ changelog.rst
@@ -21,6 +21,7 @@ Buf Fixes:
     - fixes intensity estimation in Batch widget (lower right corner) when background subtraction is activated.
     - fixes issues with representation of phases inside the contour plot of Batch widget, e.g. resetting of the
       "Show Phases" button state upon file information reloading, updating position of the phases upon user input, etc
+    - tth/q/d vertical green line position is now correctly synchronized between image, cake, pattern and batch widget
 
 0.5.3a (stable 10/10/2021)
 --------------------------